feat: add hub routing, suggest loading redesign, and notion perf caching
