+++ docs/redesign-whitepaper.md
@@ -1,0 +1,278 @@
+# DWMM 리디자인 화이트페이퍼 (기능 보존 + 디자인 전면 재작업)
+
+## 0) 문서 목적
+이 문서는 현재 DWMM 코드베이스의 기능/정보 구조를 보존하면서, **미니멀 + 고해상도 이미지 + 그라데이션 + 구조화된 레이아웃**을 핵심 축으로 한 전면 리디자인을 실행하기 위한 기준 문서다.
+
+- 범위: 시각 디자인, 내비게이션 경험, 정보 구조 표현 방식, 컴포넌트 설계 가이드.
+- 비범위: 데이터 모델 변경, 비즈니스 로직 변경, API 계약 변경.
+- 원칙: “동일한 기능을 더 좋은 감각과 더 빠른 인지로 제공한다.”
+
+---
+
+## 1) 현재 리포지토리 이해 요약
+
+### 1.1 제품 성격
+DWMM은 크게 두 축의 경험으로 구성된다.
+
+1. **북마크 큐레이션 허브 (`/`)**
+   - Supabase 기반으로 디자인 리소스 리스트를 조회/필터/검색/정렬/무한 스크롤.
+   - 카테고리/태그 필터, 추천(투표수) 정렬, 검색, 추천 사이트 제안(Suggest) 모달.
+   - 사용자 인증(익명/Google)과 투표·클릭 카운트 반영.
+
+2. **Works 포트폴리오 (`/works`, `/works/[slug]`)**
+   - Notion을 CMS로 사용해 작업물 목록/상세를 렌더링.
+   - 상세 페이지는 블록 단위 콘텐츠를 점진 로딩(커서 기반 pagination)하며 이미지/표/컬럼/리스트 등 다양한 블록을 렌더링.
+
+### 1.2 기술 스택/구조
+- **프레임워크**: Next.js (pages router 중심)
+- **데이터 소스**: Supabase(북마크/투표/인증), Notion API(Works/커리어)
+- **스타일링**: SCSS 모듈형 partial 구조
+- **실시간 인터랙션**: Supabase Realtime 기반 커서 오버레이
+
+### 1.3 핵심 사용자 플로우
+- 방문자 → 북마크 탐색(검색/필터/정렬) → 상세 링크 이동 → 투표/클릭
+- 방문자 → Works 목록 탐색 → 상세 콘텐츠 열람(이미지 중심)
+- 방문자 → Suggest 폼으로 리소스 제안
+
+---
+
+## 2) 기능 보존 원칙 (Functional Invariants)
+리디자인 시 아래 항목은 동작/의미가 동일해야 한다.
+
+1. **북마크 목록/필터/정렬/검색/무한스크롤 동작 동일**
+2. **카테고리·태그 선택 상태 유지 규칙 동일**
+3. **투표 및 클릭 카운트 노출 의미 동일**
+4. **Suggest 모달 및 제출 플로우 동일**
+5. **Works 목록/상세 진입 경로 동일**
+6. **Notion 블록 렌더링 범위(이미지, 표, 컬럼, 리스트 등) 유지**
+7. **모바일/데스크톱 모두 탐색 가능성 유지**
+
+---
+
+## 3) 디자인 비전
+
+### 3.1 비전 문장
+> “지식 탐색이 아니라 **큐레이션 전시**를 경험하게 한다. 정보는 덜어내고, 맥락은 선명하게, 이동은 유려하게.”
+
+### 3.2 감성 키워드
+- Minimal, Editorial, Luminous, Spatial, Precise
+
+### 3.3 브랜드 표현 방향
+- 타이포는 단호하고 절제되게
+- 이미지는 크게, 여백은 넓게
+- 색은 소수(뉴트럴 + 포인트 그라데이션)로 통제
+- 인터랙션은 “빠르되 과시하지 않게”
+
+---
+
+## 4) 디자인 목표 (측정 가능한 목표)
+
+1. **탐색 속도 향상**
+   - 첫 스크린에서 “무엇을 할 수 있는지”를 3초 내 파악.
+2. **콘텐츠 집중도 향상**
+   - 카드/상세에서 시선이 타이틀→요약→행동으로 자연 이동.
+3. **내비게이션 신선도 확보**
+   - 기존 필터 경험을 “구조화된 네비게이션 캔버스”로 재구성.
+4. **시각 완성도 상향**
+   - 고해상도 이미지와 그라데이션 조합을 통해 프리미엄 무드 구현.
+5. **반응형 일관성 확보**
+   - 모바일에서도 정보 밀도와 조작성 품질 유지.
+
+---
+
+## 5) 새로운 내비게이션 컨셉: Structured Discovery Canvas
+
+### 5.1 컨셉 설명
+기존 좌측 LNB + 우측 콘텐츠 구조를 확장해, 사용자가 **필터/정렬/검색 상태를 한눈에 인지**하도록 “탐색 캔버스”로 설계한다.
+
+- 상단: 글로벌 내비게이션 + 컨텍스트 타이틀
+- 좌측(데스크톱): 필터 레일(카테고리/태그/정렬)
+- 중앙: 카드 그리드(핵심 콘텐츠)
+- 우측(선택): 활성 필터 요약/추천 컬렉션(Phase 2)
+- 모바일: 하단 시트 기반 필터 허브 + 상단 고정 검색
+
+### 5.2 핵심 UX 장치
+- **Command-like Search Bar**: “검색 + 빠른 필터 토글” 통합
+- **Active Filter Chips Row**: 현재 상태 즉시 해제 가능
+- **Gradient Section Anchors**: 섹션 경계에 미묘한 그라데이션으로 시각적 리듬 형성
+- **Progressive Density**: 첫 화면은 적은 정보, 스크롤 시 메타 정보 점증
+
+---
+
+## 6) 비주얼 시스템 가이드라인
+
+### 6.1 컬러
+- Base: Neutral 900/700/500/200/50
+- Accent Gradient 1: Indigo → Cyan
+- Accent Gradient 2: Violet → Rose
+- 사용 원칙:
+  - 텍스트 대비 WCAG AA 이상
+  - 그라데이션은 배경/강조 선(Stroke)/호버 오라에 한정
+  - 카드 본문 배경은 솔리드로 가독성 우선
+
+### 6.2 타이포그래피
+- Display(히어로/헤드라인): 세리프 또는 하이콘트라스트 산세리프 (브랜드 톤 확정 필요)
+- Body/UI: 가독성 우선 산세리프
+- 스케일(예시): 56 / 40 / 28 / 22 / 18 / 16 / 14 / 12
+- 줄 길이: 본문 60~75자 기준
+
+### 6.3 레이아웃
+- 12-column grid(데스크톱), 4/6-column(태블릿), 2/4-column(모바일)
+- 주요 섹션 간 vertical rhythm: 64~120px
+- 카드 내부 간격 토큰: 8/12/16/24/32
+
+### 6.4 이미지
+- Works: 16:9 기본, 필요 시 4:3 보조
+- 북마크 카드 썸네일도 동일 비율 계열로 통일해 시각 노이즈 감소
+- lazy-load + skeleton 유지
+
+### 6.5 모션
+- 기본 전환 180~260ms, ease-out
+- 페이지 진입 시 과도한 parallax 금지
+- hover는 scale보다 **elevation/gradient edge** 중심
+
+---
+
+## 7) 화면별 구체 지시사항
+
+### 7.1 홈/북마크 (`/`)
+1. **헤더 재설계**
+   - 로고 좌측 고정, 우측 액션(로그인/로그아웃, Suggest) 정렬 체계화.
+   - 스크롤 시 컴팩트 헤더로 전환(높이 축소 + 배경 블러).
+2. **LNB → Discovery Rail**
+   - 카테고리/태그/정렬을 카드형 섹션으로 분리.
+   - 선택 상태를 색 + 아이콘 + 칩으로 중복 표기.
+3. **콘텐츠 카드**
+   - 제목/설명/메타(카테고리, 클릭수, 투표수) 계층 정리.
+   - 태그 칩은 1차 정보가 아니라 보조 정보로 시각 약화.
+4. **무결과 상태(Empty State)**
+   - 감성 카피 + 대체 행동(Suggest) CTA를 더 분명히.
+
+### 7.2 Works 목록 (`/works`)
+1. **Hero 영역**
+   - 본문 카피는 간결하게, 커리어 타임라인은 스캔 가능한 스택형.
+2. **Work Card**
+   - 대형 썸네일 + 요약 최소화(2~3줄)
+   - 연도/카테고리 메타는 상단 라인에서 통합.
+3. **무한 로딩 구간**
+   - skeleton을 카드 구조와 동일하게 유지해 레이아웃 점프 최소화.
+
+### 7.3 Works 상세 (`/works/[slug]`)
+1. **상단 메타 패널**
+   - 역할/기간/스킬/팀 정보를 한 줄씩 고정 폭 라벨로 정리.
+2. **Notion 블록 렌더링 스타일**
+   - 이미지 row/표/컬럼에 대한 스타일 토큰 통일.
+   - 본문 너비와 이미지 확장 너비를 분리(읽기/몰입 균형).
+3. **외부 링크 CTA**
+   - 페이지 상단과 하단 중복 배치, 시각 계층 통일.
+
+---
+
+## 8) 구현 가이드 (엔지니어링 관점)
+
+### 8.1 구조 전략
+- 기존 비즈니스 로직은 유지하고, **프레젠테이션 레이어 우선 분리**.
+- 권장:
+  1) 디자인 토큰(색/간격/타입/모션) 정의
+  2) 공통 컴포넌트(Button/Chip/Card/Shell) 선제 구축
+  3) 페이지별 스킨 적용
+
+### 8.2 단계별 마이그레이션
+- Phase 1: 토큰 + 레이아웃 셸 + 헤더/푸터
+- Phase 2: 북마크 페이지 재스킨
+- Phase 3: Works 목록/상세 재스킨
+- Phase 4: 인터랙션 폴리싱 + 접근성 점검
+
+### 8.3 성능/품질
+- 이미지 최적화(우선순위 이미지 선별)
+- CLS 방지(명시적 aspect ratio 유지)
+- API/데이터 접근 방식은 변경 금지(기능 보존)
+
+---
+
+## 9) 테스트 시나리오 (필수)
+
+### 9.1 기능 회귀 테스트
+1. 북마크 초기 로드 후 페이지네이션 작동 여부
+2. 카테고리 단독/태그 복수/검색 조합 필터 정확도
+3. 정렬(Newest/Oldest/Recommended) 반영
+4. 로그인/로그아웃/익명 사용자 동작
+5. 투표 상태(user_has_voted) 및 클릭수 노출
+6. Suggest 모달 열기/닫기/제출 성공·실패 플로우
+7. Works 목록 로딩 + 추가 로딩 트리거
+8. Works 상세 메타 로딩 + 콘텐츠 chunk 로딩
+9. Notion 이미지/표/리스트/컬럼 렌더링
+
+### 9.2 UI/UX 품질 테스트
+1. 360px~1920px 반응형 레이아웃
+2. 키보드 탐색(탭 순서/포커스 표시)
+3. 텍스트 대비(AA)
+4. 스켈레톤/빈 상태/에러 상태 시각 일관성
+5. 호버/포커스/액티브 상태 토큰 일관성
+
+### 9.3 성능 테스트
+1. Lighthouse(Performance, Accessibility, Best Practices)
+2. 이미지 로딩/레이아웃 안정성(CLS)
+3. 장시간 스크롤 시 메모리/프레임 저하 여부
+
+---
+
+## 10) 모호성 및 확정 필요 사항 (승인 포인트)
+아래는 구현 전에 최종 결정이 필요한 항목이다.
+
+1. **브랜드 톤 선택**
+   - A안: 미니멀 테크(산세리프 중심)
+   - B안: 에디토리얼 아트(디스플레이 세리프 혼합)
+2. **다크모드 범위**
+   - 기본 다크 only vs 라이트/다크 동시 지원
+3. **고해상도 이미지 소스 정책**
+   - Works 썸네일 원본 크기 기준, 압축 파이프라인
+4. **내비게이션 실험 강도**
+   - 보수적 개선 vs 실험적 인터랙션(애니메이션/전환)
+5. **다국어 범위**
+   - 현재 한국어/영어 혼합을 유지할지, 일관 언어 정책 적용할지
+
+---
+
+## 11) 실행 체크리스트
+- [ ] 디자인 토큰 초안 확정
+- [ ] 공통 컴포넌트(버튼/칩/카드/레이아웃) 정의
+- [ ] 홈(북마크) 시안/구현
+- [ ] Works 목록/상세 시안/구현
+- [ ] 접근성/반응형/성능 점검
+- [ ] 회귀 테스트 통과
+- [ ] 배포 전 최종 QA
+
+---
+
+## 12) 결론
+DWMM의 강점은 이미 **데이터 구조와 콘텐츠 흐름**에 존재한다. 이번 리디자인은 기능을 건드리지 않고, 사용자가 체감하는 가치를 “탐색 효율 + 시각 완성도 + 브랜드 기억성”으로 끌어올리는 작업이다. 이 문서는 그 실행 기준으로 사용한다.
+
+---
+
+## 13) 2026-03-07 확정 업데이트
+
+### 13.1 라우팅 정책 A (확정)
+- `/` 는 **DWMM Hub** 로 사용한다.
+- `/bookmarks` 는 북마크 큐레이션 전용 진입점으로 분리한다.
+- `/works` 는 Works/Blog 통합 피드로 유지한다.
+
+#### Hub 정의
+- 목적: “어디서 시작해야 하는지”를 1스크린에서 명확히 안내
+- 포함 요소:
+  1. 제품 포지셔닝 한 줄
+  2. Bookmarks 카드(검색/필터 탐색 강조)
+  3. Works/Blog 카드(케이스/아티클 탐색 강조)
+  4. 글로벌 네비게이션 active 상태 표시
+
+### 13.2 Works/Blog 정보 구조 (확정)
+- Blog와 Works는 **동일한 Notion 소스, 동일한 페이지 체계**를 공유한다.
+- 구분 방식은 라우트 분리가 아니라 **카테고리/태그 메타데이터**로 처리한다.
+
+### 13.3 Command Palette 모션 강도 (확정)
+- macOS genie 감성은 **70% 강도(꽤 강한 편)**를 기준으로 한다.
+- 다만 `prefers-reduced-motion` 환경에서는 강도를 자동 완화한다.
+
+### 13.4 Suggest 로딩 톤 (확정)
+- 로딩 카피/모션 톤은 기능적 안내보다 **감성적 큐레이션 경험**을 우선한다.
+- 단계 피드백은 유지하되, 시각/문장/전환 설계를 프리미엄 무드로 재정의한다.
