+++ docs/research/2026-03-07-suggest-loading-audit.md
@@ -1,0 +1,30 @@
+# Suggest 로딩 경험 현황 조사 (2026-03-07)
+
+## 현재 구현 위치
+- `components/bookmark/WebsiteRequestForm.js`
+- `components/bookmark/LoadingOverlay.js`
+- `styles/components/_loading.scss`
+
+## 현재 상태 요약
+1. 이미 6단계 텍스트/아이콘 기반 단계형 피드백 존재
+2. 단계 진입은 시간 기반(setTimeout)으로 고정, 실제 네트워크 상태와 느슨하게 연동
+3. 비주얼은 흰 배경 + 아이콘/텍스트 페이드 중심으로 단조롭고 감정적 보상 낮음
+4. 성공 전환은 감사 메시지로 연결되지만 로딩 구간의 스토리텔링이 약함
+5. 접근성 측면에서 현재 단계 변화에 대한 live announce가 부족
+
+## 문제 정의
+- 정보는 있지만 체험이 약함: “진행 중”이라는 사실만 전달
+- 브랜드 방향(미니멀/그라데이션/구조화/감성적)과 미스매치
+- 사용자가 기다림을 가치 있게 느끼는 장치 부족
+
+## 개선 방향
+- 감성적 문장 + 큐레이션 서사(발견→분류→큐 전달)
+- 단계 타임라인과 진행바 동시 제공
+- 글래스/그라데이션 오라로 프리미엄 무드 강화
+- 성공 직전 단계에서 기대감 연출(마이크로 모션)
+- aria-live polite로 현재 단계 방송
+
+## 반영 우선순위
+1. 로딩 오버레이 시각/카피 개편
+2. 단계 상태 표현 개선(완료/현재/대기)
+3. 접근성 및 reduced-motion 대응
