+++ lib/notion.js
@@ -2,31 +2,80 @@ import { Client } from '@notionhq/client';
 
 const notion = new Client({ auth: process.env.NOTION_API_KEY });
 
+const memoryCache = new Map();
+const inflightRequests = new Map();
+
+const getCached = (key) => {
+  const entry = memoryCache.get(key);
+  if (!entry) return null;
+  if (entry.expiresAt <= Date.now()) {
+    memoryCache.delete(key);
+    return null;
+  }
+  return entry.value;
+};
+
+const setCached = (key, value, ttlMs = 1000 * 60 * 5) => {
+  memoryCache.set(key, {
+    value,
+    expiresAt: Date.now() + ttlMs,
+  });
+  return value;
+};
+
+const withRequestCache = async (key, ttlMs, factory) => {
+  const cached = getCached(key);
+  if (cached) return cached;
+
+  if (inflightRequests.has(key)) {
+    return inflightRequests.get(key);
+  }
+
+  const pending = (async () => {
+    const value = await factory();
+    return setCached(key, value, ttlMs);
+  })();
+
+  inflightRequests.set(key, pending);
+
+  try {
+    return await pending;
+  } finally {
+    inflightRequests.delete(key);
+  }
+};
+
+
 const fetchBlockChildren = async (blockId) => {
-  let results = [];
-  let cursor = undefined;
+  return withRequestCache(`block:all:${blockId}`, 1000 * 60 * 10, async () => {
+    let results = [];
+    let cursor = undefined;
 
-  do {
-    const response = await notion.blocks.children.list({
-      block_id: blockId,
-      start_cursor: cursor,
-      page_size: 100,
-    });
-    results = results.concat(response.results);
-    cursor = response.has_more ? response.next_cursor : undefined;
-  } while (cursor);
+    do {
+      const response = await notion.blocks.children.list({
+        block_id: blockId,
+        start_cursor: cursor,
+        page_size: 100,
+      });
+      results = results.concat(response.results);
+      cursor = response.has_more ? response.next_cursor : undefined;
+    } while (cursor);
 
-  return results;
+    return results;
+  });
 };
 
 const fetchBlockChildrenPage = async (blockId, cursor, pageSize = 20) => {
-  const response = await notion.blocks.children.list({
-    block_id: blockId,
-    start_cursor: cursor || undefined,
-    page_size: pageSize,
-  });
+  const cacheKey = `block:page:${blockId}:${cursor || 'start'}:${pageSize}`;
+  return withRequestCache(cacheKey, 1000 * 60 * 5, async () => {
+    const response = await notion.blocks.children.list({
+      block_id: blockId,
+      start_cursor: cursor || undefined,
+      page_size: pageSize,
+    });
 
-  return response;
+    return response;
+  });
 };
 
 const normalizeBlocks = (blocks = []) => {
@@ -87,28 +136,34 @@ const normalizeBlocks = (blocks = []) => {
 
 const getBlocksWithChildren = async (blockId) => {
   const blocks = await fetchBlockChildren(blockId);
-  const hydrated = [];
-
-  for (const block of blocks) {
-    if (block.has_children) {
-      block.children = await getBlocksWithChildren(block.id);
-    }
-    hydrated.push(block);
-  }
+  const hydrated = await Promise.all(
+    blocks.map(async (block) => {
+      if (block.has_children) {
+        return {
+          ...block,
+          children: await getBlocksWithChildren(block.id),
+        };
+      }
+      return block;
+    })
+  );
 
   return normalizeBlocks(hydrated);
 };
 
 const getBlocksPageWithChildren = async (blockId, cursor, pageSize) => {
   const response = await fetchBlockChildrenPage(blockId, cursor, pageSize);
-  const hydrated = [];
-
-  for (const block of response.results) {
-    if (block.has_children) {
-      block.children = await getBlocksWithChildren(block.id);
-    }
-    hydrated.push(block);
-  }
+  const hydrated = await Promise.all(
+    response.results.map(async (block) => {
+      if (block.has_children) {
+        return {
+          ...block,
+          children: await getBlocksWithChildren(block.id),
+        };
+      }
+      return block;
+    })
+  );
 
   return {
     content: normalizeBlocks(hydrated),
@@ -410,66 +465,65 @@ const getYearFromProperties = (properties, createdTime) => {
 
 export const getPublishedPosts = async () => {
   const dynamicDataSourceId = await getWorksDataSourceId();
-
-  // dataSources.query를 사용하고, data_source_id를 전달
-  const response = await notion.dataSources.query({
-    data_source_id: dynamicDataSourceId,
-    // filter: {
-    //   property: 'status',
-    //   select: {
-    //     equals: 'published',
-    //   },
-    // },
-    sorts: [
-      {
-        property: 'publishedAt',
-        direction: 'descending',
-      },
-    ],
+  return withRequestCache(`works:all:${dynamicDataSourceId}`, 1000 * 60 * 2, async () => {
+    const response = await notion.dataSources.query({
+      data_source_id: dynamicDataSourceId,
+      sorts: [
+        {
+          property: 'publishedAt',
+          direction: 'descending',
+        },
+      ],
+    });
+    return response.results;
   });
-  return response.results;
 };
 
 export const getPublishedPostsPage = async ({ cursor = null, pageSize = 5 } = {}) => {
   const dynamicDataSourceId = await getWorksDataSourceId();
+  const cacheKey = `works:page:${dynamicDataSourceId}:${cursor || 'start'}:${pageSize}`;
 
-  const response = await notion.dataSources.query({
-    data_source_id: dynamicDataSourceId,
-    start_cursor: cursor || undefined,
-    page_size: pageSize,
-    sorts: [
-      {
-        property: 'publishedAt',
-        direction: 'descending',
-      },
-    ],
+  return withRequestCache(cacheKey, 1000 * 60 * 2, async () => {
+    const response = await notion.dataSources.query({
+      data_source_id: dynamicDataSourceId,
+      start_cursor: cursor || undefined,
+      page_size: pageSize,
+      sorts: [
+        {
+          property: 'publishedAt',
+          direction: 'descending',
+        },
+      ],
+    });
+
+    return response;
   });
-
-  return response;
 };
 
 export const getPostBySlug = async (slug) => {
   const dynamicDataSourceId = await getWorksDataSourceId();
 
-  // dataSources.query를 사용하고, data_source_id를 전달
-  const response = await notion.dataSources.query({
-    data_source_id: dynamicDataSourceId,
-    filter: {
-      property: 'slug',
-      rich_text: {
-        equals: slug,
+  return withRequestCache(`works:slug:${dynamicDataSourceId}:${slug}`, 1000 * 60 * 10, async () => {
+    const response = await notion.dataSources.query({
+      data_source_id: dynamicDataSourceId,
+      filter: {
+        property: 'slug',
+        rich_text: {
+          equals: slug,
+        },
       },
-    },
+      page_size: 1,
+    });
+    return response.results[0];
   });
-  return response.results[0];
 };
 
 export const getCareerEntries = async () => {
   const dataSourceId = await getCareerDataSourceId();
-  const response = await notion.dataSources.query({
+  const response = await withRequestCache(`career:list:${dataSourceId}`, 1000 * 60 * 10, async () => notion.dataSources.query({
     data_source_id: dataSourceId,
     page_size: 100,
-  });
+  }));
 
   const entries = response.results.map((page) => {
     const properties = page.properties || {};
