+++ pages/api/works/list.js
@@ -47,6 +47,8 @@ export default async function handler(req, res) {
     return res.status(200).json(cached.data);
   }
 
+  res.setHeader('Cache-Control', 'public, s-maxage=120, stale-while-revalidate=600');
+
   try {
     const response = await getPublishedPostsPage({ cursor, pageSize });
     const works = transformNotionData(response.results || []);
