+++ pages/api/works/meta.js
@@ -79,6 +79,8 @@ export default async function handler(req, res) {
     return res.status(200).json(cached.data);
   }
 
+  res.setHeader('Cache-Control', 'public, s-maxage=120, stale-while-revalidate=600');
+
   try {
     const postData = await getPostBySlug(slug);
     const post = transformPostData(postData);
