+++ styles/layout/_header.scss
@@ -53,3 +53,20 @@
     }
   }
 }
+
+.header-link.is-active {
+  color: $grey-12;
+  font-weight: 600;
+  position: relative;
+}
+
+.header-link.is-active::after {
+  content: '';
+  position: absolute;
+  left: 8px;
+  right: 8px;
+  bottom: 2px;
+  height: 2px;
+  background: linear-gradient(90deg, #60a5fa, #818cf8, #f472b6);
+  border-radius: 999px;
+}
