+++ styles/pages/_home.scss
@@ -1,11 +1,91 @@
-// _home.scss
 @use '../utils/variables' as *;
 @use '../utils/mixins' as *;
 @use 'sass:color';
 
 .home {
-    @include flex-center;
-    height: 100vh;
-    background-color: color.adjust($black-lighter, $lightness: 40%);
+  @include flex-center;
+  height: 100vh;
+  background-color: color.adjust($black-lighter, $lightness: 40%);
+}
+
+.dwmm-hub {
+  min-height: calc(100vh - 130px);
+  background: radial-gradient(circle at 18% 10%, #f1f5ff 0%, #f8f6ff 45%, #f7f2ef 100%);
+  padding: 4.5rem clamp(1.2rem, 3vw, 2.5rem) 5rem;
+}
+
+.dwmm-hub__hero {
+  max-width: 860px;
+  margin: 0 auto 2.2rem;
+
+  h1 {
+    margin: 0;
+    font-size: clamp(2rem, 5vw, 3.4rem);
+    line-height: 1.15;
+    letter-spacing: -0.02em;
   }
-  
+}
+
+.dwmm-hub__eyebrow {
+  text-transform: uppercase;
+  letter-spacing: 0.13em;
+  font-size: 0.72rem;
+  color: #6b7280;
+  margin-bottom: 0.9rem;
+}
+
+.dwmm-hub__subtitle {
+  margin-top: 0.9rem;
+  max-width: 620px;
+  line-height: 1.6;
+  color: #4b5563;
+}
+
+.dwmm-hub__grid {
+  max-width: 860px;
+  margin: 0 auto;
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 1rem;
+}
+
+.dwmm-hub__card {
+  text-decoration: none;
+  color: inherit;
+  border-radius: 20px;
+  border: 1px solid rgba(17, 24, 39, 0.08);
+  background: linear-gradient(165deg, rgba(255, 255, 255, 0.92), rgba(255, 255, 255, 0.65));
+  backdrop-filter: blur(12px);
+  padding: 1.2rem;
+  transition: transform 260ms cubic-bezier(.22,1,.36,1), box-shadow 260ms ease;
+
+  &:hover {
+    transform: translateY(-2px) scale(1.01);
+    box-shadow: 0 12px 28px rgba(55, 65, 81, 0.14);
+  }
+
+  h2 {
+    margin: 0.28rem 0 0.6rem;
+    font-size: 1.25rem;
+  }
+
+  p {
+    margin: 0;
+    font-size: 0.95rem;
+    color: #4b5563;
+    line-height: 1.55;
+  }
+}
+
+.dwmm-hub__card-kicker {
+  font-size: 0.72rem;
+  text-transform: uppercase;
+  letter-spacing: 0.11em;
+  color: #6b7280;
+}
+
+@media (max-width: 760px) {
+  .dwmm-hub__grid {
+    grid-template-columns: 1fr;
+  }
+}
